[kotlin] reimplement one line method check
Previously we assumed that a method is located on a single line if the first location in this method was equal to the last location. However, this assumption may sometimes be wrong, which can lead to skipping whole methods during a step into. This commit reimplements `isOneLineMethod` by checking that all locations in a method have the same line number, except for fake locations and locations that were inlined.

(cherry picked from commit 1eacae992f3a6972b259e88b94f94b4646eb43c2)

IJ-MR-23417

GitOrigin-RevId: f9aff218908517c8760b92599ba17cca0d5b440f