Load JVM target 1.6 as 1.8 when importing Maven projects
JVM target 1.6 is going to be removed in Kotlin 1.7 (KT-45165), but we'd
like to avoid breaking local compilation for users with old Kotlin
version in their Maven project, but a new Kotlin IDE plugin version. So
we import it as the minimal supported JVM target, 1.8. Since it
can influence behavior or project tests, we display notification on each
import, to nudge the user into migrating the project to JVM target 1.8.

Since JPS plugin is unbundled from the Kotlin IDE plugin starting from
222, this change needs to go to {221, 213, 212} x {master, 1.7.0}.

 #KTIJ-21515 Fixed

Merge-request: IJ-MR-23324
Merged-by: Alexander Udalov <Alexander.Udalov@jetbrains.com>
(cherry picked from commit cd9835cc586002ca0ac5cfec35f9bd2d83543a7e)

See the next commit to understand reasons for the cherry-pick into 222
release branch. KTIJ-21954

See the commit after the next to understand why the commit is actually
required not only for `222` but for `master` too.

GitOrigin-RevId: 3173e2ae64aa447dbe58a161f2109ab2d468877a