Move KotlinFacetSettingsProvider from kotlin repo to intellij repo
This class isn't used in kotlin repository (kotlin/jps to be precise) and
the class was causing troubles in kotlin/jps (I was dropping dependency
on intellij-core in kotlin/jps and KotlinFacetSettingsProvider depends on
com.intellij.openapi.module.Module and
com.intellij.openapi.project.Project which are intellij-core classes)

The class was moved from
`jps/jps-common/src/org/jetbrains/kotlin/config/KotlinFacetSettings.kt`
file (so you can find a commit in which it was moved)

Also, it looks like KotlinFacetSettingsProvider interface isn't needed
because there is only one implementation, and it doesn't hold any state
so it could be converted from a service to simple a class. But,
unfortunatelly, at least two external plugins reference this class, so I
decided not to break compatibility for nothing.

The list of plugins that reference KotlinFacetSettings:
* org.spekframework
* melabsinthiatum.kotlin-mpp-shared-items-browser

GitOrigin-RevId: ab9fe1205e5f95c30485e54cb1b1b059b4f89a01