[Kotlin][Import] Deprecate error prone 'allModules' assertion function
allModules is deprecated, because it is very easy to misunderstand:
The assertions will only run for all modules that are explicitly asserted
*after* this function call. If no module gets additional assertions, this
method is effectively useless. If the modules are declared before this method call,
then it is effectively useless.

As an alternative a new 'forEachModule' function is implemented.

KTIJ-21692

GitOrigin-RevId: 395cf9ec1333bee18d3d6af7af641b55f000b41b