[kotlin] Ensure all Kotlin modules are present in meta-modules
There are several meta-modules:
- 'kotlin.all-tests' and 'kotlin.fir.all-tests'. Used as a test classpath on CI.
- 'kotlin.plugin' and 'kotlin.plugin-fir'. Used mainly for XML plugin descriptor resolution.
- 'intellij.kotlin.plugin.community.main' and 'intellij.kotlin.plugin.community.fir.main'. Used for running IDEA from sources.

All relevant Kotlin IDE plugin modules should be present in each of these meta-modules. In addition, there is 'KotlinPluginBuilder', a Groovy script which assembles plugin artifact JARs.

It is easy to forget putting newly created modules in each of these locations, so here is the consistency test.

GitOrigin-RevId: 990dfb1c5447e22ab89fb4aeb2cd3d8e937aa110