Cleanup: Drop feature flag for "unbundled Kotlin JPS" feature
"unbundled Kotlin JPS" feature (KTIJ-11633) has been enabled by default
for more than a month already and we didn't face major issues with it
(Well, KTIJ-22098 is quite bad but it seems that it affects not a lot of
users).

Drop the feature flag to cleanup the code.

The only code logic behaviour this commit brings is that we treat
"unbundled JPS" feature enabled for "default project" (Welcome screen)
too. But it shouldn't make any user visible difference.

GitOrigin-RevId: 1c3d45c404f73e9b29fef8c976221616da866050