[Kotlin] Fix Gradle configurator tests broken by e0cec6ff
In e0cec6ff, we removed one of no-more-supported Kotlin compiler features
from these Gradle tests. However, the main meaning of these tests was
to add another compiler argument in a situation when some feature
is already available. So in this commit I simply replaced the deleted feature
with some other feature.

GitOrigin-RevId: 66acccdbf4a412231051e7792af63dab734a3097