[kotlin] fix evaluation of captured variables in SAM lambdas
In the old Kotlin backend captured variables are stored as fields of a generated lambda class. When evaluating these variables, the evaluator searched for the corresponding fields in `this` reference of a lambda class. Captured variables were also "lifted" out of `this` in the variables view. For example if `foo` was captured, a user will observe it as `foo` instead of `this.$foo`. However, in the IR backend SAM lambdas are generated as functions, and captured variables are stored in the LVT table. The evaluator tried searching for them in fields of `this`, when there wasn't any `this` on a frame. A user also observed captured variables with a dollar prefix (e.g. `$foo`) which was confusing. This commit fixes the evaluator issue and drops the dollar prefix for captured variables.

^KTIJ-22654

GitOrigin-RevId: b3ef9359231cc6ab8954f7b491c997583ccc660f