[kotlin] Generate JUnit 5 test classes and functions as public, not internal
Add option to J2K to generate declarations as public instead of internal.

The actions to generate test classes/functions are still using the old J2K,
so the option is supported both in old and new J2K. When the actions are migrated
to use the new J2K, hopefully it should just continue to work.

^KTIJ-11681 Fixed

GitOrigin-RevId: 1e7211e6981ab6a281b4b957b4e60cf85fdf3856