[kotlin] J2K: render one enum entry per line
Currently, we don't preserve the original line breaks between enum entries.
But the default of multiple entries per line is problematic, because
it can result in weird formatting that is time-consuming to correct.

Seems like it's better to render one enum entry per line: it won't create
these problems and is easier to manually correct if needed.

^KTIJ-23382 Fixed

GitOrigin-RevId: cc9512acada411d3025c6972dcf65e8acbb1f237