[kotlin] KTIJ-23574 Scripts: don't show script libraries at project settings
Prior to this commit we utilized platform `LibraryEntity` to represent
script libraries. As a benefit we got indexing and settings UI support
out of the box.
Not to show script libraries effectively means transition to a new entity,
the platform is unaware of. `KotlinScriptLibraryEntity` is the very same one.
It's simpler than its predecessor which also affects indexing.

`KotlinScriptLibraryIdIteratorBuilder`,
`KotlinScriptLibraryIndexableIteratorHandler`,
`KotlinScriptLibraryIndexableFilesIteratorImpl`

are the price we pay introducing the custom entity and organizing its indexing
ourselves.

^KTIJ-23574 fixed

GitOrigin-RevId: e8da6e30083381f12239a7c520f39be4f4293f68