[kotlin] fix exception invoking generate test action
The RemoveUnnecessaryParenthesesIntention class was placed into the module with the new plugin model.
J2KPostProcessingRegistrarImpl from the old J2K (which is used for the test generation) uses RemoveUnnecessaryParenthesesIntention
which violated the contract that modules which are not migrated to the new plugin model should not use functionality from the modules which were migrated.

This commit extracts J2KPostProcessingRegistrarImpl to a new-plugin style module to solve this issue.

^KTIJ-23593 fixed

GitOrigin-RevId: 2ac79947fa9afd231cae021fc2dc7f9ea633935a