Implement InsertExplicitTypeArgumentsIntention in K2.
A short implementation of an intention to add explicit type arguments to
a function call. Note that the K1 version of this intention is also used
by some other intentions and refactorings - this change does _not_
attempt to port those usages.

Also adds tests for ^KTIJ-13038, to ensure that issue does not occur in
the K2 implementation. The test is currently ignored for K1, but can be
restored when KTIJ-13038 is fixed.

^KTIJ-22989 Fixed

GitOrigin-RevId: 502664d84fa3b140f271ea34de44de2c881c421b