[kt, tests] Rework handling of KGP_VERSION placeholder in testdata
Previously, the code used to substitute the placeholders with actual version
number in testdata before comparing it to the actual file.

This doesn't work that well with FileComparisonException in IDE, as you'll
see the diff with actual versions, and applying diff would write actual
version numbers into the testdata file.

Instead, now it's vice-versa: the code doesn't touch expected file, but
folds actual version numbers in actual-text back into {{KGP_VERSION}}
placeholder.

This way, we still guarantee that during any specific runs versions were
exactly those that we expected (otherwise, 'replace'-call wouldn't have
matched it), but provide better UX in IDE (the diff will contain only
placeholders, so it's easier to read and you can apply it to testdata)

GitOrigin-RevId: 6f94268d8c8506e0019a7a534f47e0da05a711db