KTIJ-23034 Fix not captured outer 'this' in indy lambdas when evaluating expression
In generated code of lambda outer 'this' is named 'this$0', and because of this is filtered out of visibleVariables in StackFrame and can't be found in VariableFinder. But we can still find it in argumentValues.

(cherry picked from commit 9d296c7224a96ae1e59524c6efb2cff1c69517e1)

IJ-MR-102042

GitOrigin-RevId: 9cecec4f4da086feb74d6a01cc2856316a41190c