KTIJ-23034 Fix incorrect behaviour when breakpoint is on line with several indy lambdas
When several indy lambdas (one of which is non-SAM) are on same line,
and breakpoint is set, this leads to different bugs in debugger
(see enabled tests). This is because KotlinPositionManager incorrectly
calculated SourcePosition for Location, because it expected that all
indy lambdas are SAM, but this is not true, when -Xlambdas=indy is set.

(cherry picked from commit 3cfa415a757136cf31077377bc16cdfd0e15c3f0)

IJ-MR-102042

GitOrigin-RevId: a9f9489dab12171e69e28f18c830696965dd54a2