[IDEA-291565] ExternalProject: Import Gradle's identityPath and path
The identityPath will already contain composite build prefixes
and can safely be used to construct paths for corresponding tasks

This commit uses ProjectInternal to retrieve the identityPath.
This is a midterm solution agreed upon with Gradle.
The API will later be marked as 'used by IDEA' on Gradle side.

A new proper public tooling API is proposed that can be used later.

^KTIJ-24771 Verification Pending
^IDEA-291565 fixed

GitOrigin-RevId: 21349a1023c715d0f3dbac4e8932d2cb9c92a87b