[kotlin] Erase type params when compiling `generated_for_debugger_fun`
https://github.com/JetBrains/kotlin/commit/c447b91101e4c72058dede19bfc8d6aafea16385
broke expression evaluation because of a new assertion in the mangler verifying
that all type parameters that the mangler encounters must have their containers
in scope.

When we generated parameters for the `generated_for_debugger_fun` method,
we were just copying the types of captured values, thus possibly ending up with
those types being type parameters that were now used outside the scope they
were defined in, making the Kotlin compiler upset.

Now we check if a parameter type is a type parameter, and if that's the case,
we use its upper bound for the method parameter type.

GitOrigin-RevId: 7fc964644a9e0d699b5d7bf97ecdb432d534fa3e