[kotlin] add progress indicator when searching for inline call sites
Previously when trying to place a breakpoint in an inline function, the debugger would have to find all of its call sites to perform this action. If you, for example, put a breakpoint in one of standard library functions, the searching would take a huge amount of time. Things became worse, when you wanted to unset a breakpoint and the searching process was still not finished. Unsetting a breakpoint required a write action, which waited for the previously launched read action to find all the inline function call sites. As a result freeze could happen and the debugger UI would be unresponsive until all of the call sites were found. This commit replaces EmptyProgressIndicator with ProgressIndicatorBase, which allows the action to be cancelled when needed.

^KTIJ-24736 Fixed

GitOrigin-RevId: 9e2f2b0538ba5ef2fa035ea5df6e7e10f8405f73