[debugger] Change IDE compilation to CLI in tests [1/4]
This series of commits changes the compilation in debugger tests from
IDE to CLI.

Why this is needed:

1. For some reason the compilation in IDE doesn't do a lot of checks.
That's why even despite analysisResult.throwIfError() there were a lot
of undetected compile errors in the tests.

2. The `GenerationUtils.generateFiles` used now doesn't
support K2 compilation (does not look at the
CommonConfigurationKeys.USE_FIR). `GenerationUtils.compileFiles`
supports it, but it couldn't be adapted to the tests in
IDEA because it is highly tied to the Kotlin repository tests,
and it starts throwing errors in a lot of places in case of
IDEA infrastructure. Rewriting to CLI is more straightforward here.

3. Compiling via CLI (KotlinCompilerStandalone) is used in a lot
of other tests in the idea. Compiling in IDE (using GenerationUtils)
is only used in debugger tests and `PositionManagerTest`.

4. The debugger tests already have a CLI compilation mode, which is
enabled for MPP tests. Removing the IDE compilation mode will
simplify the code.

This change is divided into several commits for convenience.

In this commit:

- removed compilation in IDE for all tests except LowLevelDebuggerTest
- fixed the compilation options in the CLI to allow for simultaneous
compilation of .kt and .java or .kts
- found compilation errors in tests were fixed
- added DebuggerTestCompilerFacilityTest for compilation errors reporting
- Arguments from // COMPILER_ARGUMENTS directive added to
DebuggerTestCompilerFacility. Otherwise tests `smartStepWithInlineClass`
and `filterFunctionCallsFromInlineClass` won't compile with -Xuse-old-backend

See KTIJ-24739

(cherry picked from commit e67259cb577b0d6836b9f0d1a79dc2260b2258a2)

IJ-MR-105785

GitOrigin-RevId: 7b1c810db9df4f255034367bec33942768025c30