KTIJ-25125 [kotlin] Optimize `RemoveRedundantQualifierNameInspection` in K2 IDE
Instead of calling reference shortener on every qualified element in the file,
call it for every top-level declaration in the file (see KT-57966 for the reason
why not on the whole file).

This reduces the number of on-air resolve calls, which are quite expensive.

Mute some (questionable) test data due to the bugs in the reference shortening
implementation (KT-58044) and in the compiler (KT-58034).

^KTIJ-25125 Fixed

GitOrigin-RevId: 91328420e09d94bb79cebd0f89f8f1fba74bde1f