[KTIJ-25227] Use CompilerArgumentsAware to resolve CompilerArguments for KGP < 1.9
This commit will take care of removing the classpath of the Compile Task
from the resolved CompilerArguments as this classpath is undesired and can
take too much memory. JPS builds will restore the classpath from
the IJ project model. Classpath entries added as freeCompilerArgs are still
relevant.

This commit also ensures that the jvmTarget is present in the
List of argument Strings for older KGP plugins that omitted 1.6 targets
as it was their default value.

KTIJ-25227

GitOrigin-RevId: 721dee6ee0f490b7c06c5fd6af5e777cf5760466